+++ content_editor/static/content_editor/save_shortcut.js
@@ -1,0 +1,9 @@
+/* global django */
+django.jQuery(function ($) {
+  $(document).keydown(function handleKeys(event) {
+    if (event.which == 83 && (event.metaKey || event.ctrlKey)) {
+      $("form input[name=_continue]").click();
+      return false;
+    }
+  });
+});